+++ nion/utils/Event.py
@@ -5,6 +5,8 @@ An event object to which to attach listeners.
 # standard libraries
 import logging
 import threading
+import traceback
+import sys
 import weakref
 
 # third party libraries
@@ -16,8 +18,9 @@ import weakref
 
 class EventListener:
 
-    def __init__(self, listener_fn):
+    def __init__(self, listener_fn, trace):
         self.__listener_fn = listener_fn
+        self.tb = traceback.extract_stack() if trace else None
         # the call function is very performance critical; make it fast by using a property
         # instead of a method lookup.
         if callable(listener_fn):
@@ -57,10 +60,9 @@ class Event:
 
     def listen(self, listener_fn, *, owner=None, trace=False):
         """Add a listener function and return listener token. Token can be closed or deleted to unlisten."""
-        listener = EventListener(listener_fn)
+        listener = EventListener(listener_fn, self.__trace)
         def remove_listener(weak_listener):
             if trace:
-                import traceback
                 traceback.print_stack()
             with self.__weak_listeners_mutex:
                 self.__weak_listeners.remove(weak_listener)
@@ -76,8 +78,25 @@ class Event:
             self.__listeners[id(weak_owner)] = listener, weak_owner
         return listener
 
+    def __print_event_exception(self, exc_listener):
+        print("Event Fire Traceback (most recent call last):", file=sys.stderr)
+        etype, value, tb = sys.exc_info()
+        for line in traceback.StackSummary.from_list(traceback.extract_stack()[:-2]).format():
+            print(line, file=sys.stderr, end="")
+        if exc_listener.tb is not None:
+            print(f"Event Listener Traceback (most recent call last)", file=sys.stderr)
+            for line in traceback.StackSummary.from_list(exc_listener.tb[:-2]).format():
+                print(line, file=sys.stderr, end="")
+        print(f"Event Handler Traceback (most recent call last)", file=sys.stderr)
+        traceback_exception = traceback.TracebackException(type(value), value, tb)
+        traceback_exception.stack = traceback.StackSummary.from_list(traceback_exception.stack[1:])
+        traceback_exception.exc_traceback = None  # prevent printing of header
+        for line in traceback_exception.format(chain=True):
+            print(line, file=sys.stderr, end="")
+
     def fire(self, *args, **keywords):
         """Calls listeners (in order added) unconditionally."""
+        listener = None
         try:
             with self.__weak_listeners_mutex:
                 listeners = [weak_listener() for weak_listener in self.__weak_listeners]
@@ -85,13 +104,11 @@ class Event:
                 if listener:
                     listener.call(*args, **keywords)
         except Exception as e:
-            import traceback
-            logging.debug("Event Error: %s", e)
-            traceback.print_exc()
-            traceback.print_stack()
+            self.__print_event_exception(listener)
 
     def fire_any(self, *args, **keywords):
         """Calls listeners (in order added) until one returns True or else return False."""
+        listener = None
         try:
             with self.__weak_listeners_mutex:
                 listeners = [weak_listener() for weak_listener in self.__weak_listeners]
@@ -101,13 +118,11 @@ class Event:
                         return True
             return False
         except Exception as e:
-            import traceback
-            logging.debug("Event Error: %s", e)
-            traceback.print_exc()
-            traceback.print_stack()
+            self.__print_event_exception(listener)
 
     def fire_all(self, *args, **keywords):
         """Calls listeners (in order added) until one returns False or else return True."""
+        listener = None
         try:
             with self.__weak_listeners_mutex:
                 listeners = [weak_listener() for weak_listener in self.__weak_listeners]
@@ -117,7 +132,4 @@ class Event:
                         return False
             return True
         except Exception as e:
-            import traceback
-            logging.debug("Event Error: %s", e)
-            traceback.print_exc()
-            traceback.print_stack()
+            self.__print_event_exception(listener)